Compiler improvements:
- Move structural type reconstruction down from inferTypes into
  expandTables where a full tree transformation has to be done anyway.

- Remove the assignTypes phase. Types are now preserved everywhere.

- Some simplifications.
