+++ slick-testkit/src/main/scala/com/typesafe/slick/testkit/tests/PagingTest.scala
@@ -19,18 +19,18 @@ class PagingTest extends AsyncTest[RelationalTestDB] {
     def q5 = q1 take 5 drop 3
     val q6 = q1 take 0
 
-    seq(
-      ids.schema.create,
-      ids ++= (1 to 10),
-      q1.result.map(_ shouldBe (1 to 10).toList),
-      q2.result.map(_ shouldBe (1 to 5).toList),
-      ifCap(rcap.pagingDrop)(seq(
+    for {
+      _ <- ids.schema.create
+      _ <- ids ++= (1 to 10)
+      _ <- mark("q1", q1.result).map(_ shouldBe (1 to 10).toList)
+      _ <- mark("q2", q2.result).map(_ shouldBe (1 to 5).toList)
+      _ <- ifCap(rcap.pagingDrop)(seq(
         q3.result.map(_ shouldBe (6 to 10).toList),
         q4.result.map(_ shouldBe (6 to 8).toList),
         q5.result.map(_ shouldBe (4 to 5).toList)
-      )),
-      q6.result.map(_ shouldBe Nil)
-    )
+      ))
+      _ <- mark("q6", q6.result).map(_ shouldBe Nil)
+    } yield ()
   }
 
   def testCompiledPagination = {
